requirements.txt
absl-py==2.2.0
accelerate==1.5.2
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anthropic==0.49.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==4.0.3
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.3
bitsandbytes==0.45.3
black==21.4b2
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
clip==0.2.0
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cryptography==45.0.5
cycler==0.12.1
dashscope==1.23.8
dataclasses-json==0.6.7
datasets==2.19.1
debugpy==1.8.13
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
detectron2 @ git+https://github.com/MaureenZOU/detectron2-xyz.git@42121d75e10d9f858f3a91b6a39f5722c02868f0
diffusers==0.29.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.1
einops-exts==0.0.4
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
faiss==1.9.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.3.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gdcm==1.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-genai==1.24.0
gradio==5.23.1
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
hydra-core==1.3.2
idna==2.10
imageio==2.37.0
importlib_metadata==8.6.1
iopath==0.1.9
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
langchain==0.3.21
langchain-anthropic==0.3.10
langchain-community==0.3.20
langchain-core==0.3.47
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
latex2mathml==3.77.0
lazy_loader==0.4
lightning-utilities==0.14.2
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
-e git+https://github.com/bowang-lab/MedRAX.git@0dd6764aad5f3213622dc3717e55928b2e10ef3a#egg=medrax
mistune==3.1.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpi4py @ file:///croot/mpi4py_1671223370575/work
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.24.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch @ git+https://github.com/usuyama/open_clip.git@01a53cc46662d8b28da0c9d73271bfb8f6f4b20d
openai==1.68.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
peft==0.15.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydicom==3.0.1
pydot==3.0.4
pydub==0.25.1
Pygments==2.19.1
pygraphviz==1.14
pylibjpeg==2.0.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-fid==0.3.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
ruff==0.11.2
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.24.1
setproctitle==1.3.5
shellingham==1.5.4
shortuuid==1.0.13
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
stringzilla==3.12.3
sty==1.0.6
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
timeout-decorator==0.5.0
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
toolz==0.12.1
torch==2.6.0
torchattacks==3.5.1
torchaudio==2.6.0
torchmetrics==1.3.2
torchvision==0.21.0
torchxrayvision==1.3.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
triton==3.2.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.34.0
validators==0.35.0
wandb==0.19.8
wavedrom==2.0.3.post3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
xformers==0.0.29.post3
xlsxwriter==3.2.5
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
